+++ src/devicecontroller/NoOpDeviceController.ts
@@ -4,13 +4,18 @@
 import DeviceChangeObserver from '../devicechangeobserver/DeviceChangeObserver';
 import DeviceControllerBasedMediaStreamBroker from '../mediastreambroker/DeviceControllerBasedMediaStreamBroker';
 import NoOpMediaStreamBroker from '../mediastreambroker/NoOpMediaStreamBroker';
-import Device from './Device';
+import AudioInputDevice from './AudioInputDevice';
 import DevicePermission from './DevicePermission';
+import VideoInputDevice from './VideoInputDevice';
 import VideoQualitySettings from './VideoQualitySettings';
 
 export default class NoOpDeviceController
   extends NoOpMediaStreamBroker
   implements DeviceControllerBasedMediaStreamBroker {
+  constructor(_options?: { enableWebAudio?: boolean }) {
+    super();
+  }
+
   listAudioInputDevices(): Promise<MediaDeviceInfo[]> {
     return Promise.resolve([]);
   }
@@ -23,11 +28,11 @@ export default class NoOpDeviceController
     return Promise.resolve([]);
   }
 
-  chooseAudioInputDevice(_device: Device): Promise<DevicePermission> {
+  chooseAudioInputDevice(_device: AudioInputDevice): Promise<DevicePermission> {
     return Promise.reject();
   }
 
-  chooseVideoInputDevice(_device: Device): Promise<DevicePermission> {
+  chooseVideoInputDevice(_device: VideoInputDevice): Promise<DevicePermission> {
     return Promise.reject();
   }
 
@@ -63,6 +68,4 @@ export default class NoOpDeviceController
   getVideoInputQualitySettings(): VideoQualitySettings | null {
     return null;
   }
-
-  enableWebAudio(_flag: boolean): void {}
 }